% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories_function.R
\name{create.directories}
\alias{create.directories}
\title{Create Results Directories}
\usage{
create.directories(homedir)
}
\arguments{
\item{homedir}{The directory in which the result folders will be created.}
}
\value{
Creates storage directories.
}
\description{
Creates the directories, in which the results are going to be stored.
}
\details{
All files will be stored in the main folder "FAMoS-Results". It contains the following subdirectories:
\describe{
  \item{Figures}{Contains plots of the profiles.}
  \item{Fits}{Contains the fitted parameter values of each of the tested models.}
  \item{LogFiles}{Contains the log files of the individual model fitting runs.}
  \item{Status}{Contains the status files of the individual model fitting runs.}
  \item{Tables}{Contains the profile likelihood data frames.}
  }
}
\examples{
create.directories(homedir = getwd())
}
