% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_parameters_function.R
\name{parscale.parameters}
\alias{parscale.parameters}
\title{A Fix for Parscale in Optim}
\usage{
parscale.parameters(par, scale, fix = 1)
}
\arguments{
\item{par}{A vector containing the values of the original parameters.}

\item{scale}{A vector containing the corresponding absolute scaling values that will be used during the first steps in \code{\link{optim}}.}

\item{fix}{Integer containing the index of the parameter that will be scaled by 10\% of its original value, meaning the corresponding entry in scale will be overwritten (\code{parscale} in \code{\link{optim}} needs one value like this). Default to 1.}
}
\value{
A vector that can be used to scale \code{parscale} in \code{\link{optim}} accordingly.
}
\description{
This function is designed to use the built-in option \code{parscale} in \code{\link{optim}} with absolute scaling values.
}
\examples{
test.func<-function(x){
print(x)
3*x[1]+4*x[2]
}

pars <- c(1, 1000, 10)
scaling <- c(0.1, 3000, 10)

p.scale <- parscale.parameters(par = pars, scale = scaling)

optim(par = pars, fn = test.func, control = list(maxit = 10, parscale = p.scale, trace = TRUE))

}
