% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_and_fit_function.R
\name{unite.and.fit}
\alias{unite.and.fit}
\title{Combine Fitted and Non-fitted Parameters}
\usage{
unite.and.fit(par, no.fit, par.names, fit.fn, ...)
}
\arguments{
\item{par}{A named vector containing all parameters that are supposed to be fitted.}

\item{no.fit}{A named parameter vector containing all parameters that are not supposed to be fitted.}

\item{par.names}{The names of all parameters}

\item{fit.fn}{The cost function (see \code{\link{create.profile}} for more details).}

\item{...}{Other arguments.}
}
\value{
Returns the negative log-likelihood as calculated by the specified cost function
}
\description{
Combines fitted and non-fitted parameters and calls the fitting function. Serves as a wrapping function for the user-specified fitting function \code{fit.fn} (see \code{\link{create.profile}}).
}
\examples{
#set parameters and cost function
fit.par <- c(p1 = 2, p2 = 4)
name.par <- c("p1", "p2", "p3")
defaults <- list(p1 = 0, p2 = 2, p3 = 4)
cost.function <- function(parms){
    parms[1] + parms[2] + parms[3]
}

#call unite.and.fit
unite.and.fit(par = fit.par, no.fit = c(p3 = 1),par.names = name.par, fit.fn = cost.function)
}
