% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_profile_function.R
\name{get.profile}
\alias{get.profile}
\title{Create Profile Likelihood Data Frame}
\usage{
get.profile(which.par, range, homedir = getwd(), wait = FALSE)
}
\arguments{
\item{which.par}{The names of the parameter that the profile should be returned for.}

\item{range}{A numeric vector determining the values for which the profile should be calculated.}

\item{homedir}{The directory in which the folder \code{"Profile-Results"} is found. Default to \code{\link{getwd}}().}

\item{wait}{Logical. If true, \code{get.profile} will wait if a certain profile value has not been fitted yet (i.e. the corresponding status file is not equal to "done").}
}
\value{
A data frame containing the negative log-likelihood and the fitted parameter values for a specific range created by \code{\link{create.profile}}.
}
\description{
Given previously calculated profile values by \code{\link{create.profile}}, \code{get.profile} merges the results together into a data frame.
}
\examples{
#define cost function
cost_function <- function(parms){
y <- parms[1] + parms[2]*c(1:3)
LL <- sum((y - c(1:3))^2)
}

#create profile values
create.profile(which.par = "p1",
               par.names = c(p1 = 1, p2 = 3),
               range = list(seq(0,5,1)),
               fit.fn = cost_function,
               future.off = TRUE)

#retrieve the calculated profile
res <- get.profile(which.par = "p1", range = seq(0,5,1))
}
